Add files via upload
turtleStart.sh needs to be placed in the following location on the Jetson: /home/ubuntu/jetsonbot/devel. rc.local needs to be placed within /etc/. Be sure to merge any pre-existing content in rc.local with what in the attached. 

turtleStart.sh may need to have executable permissions added to it in order for it to run. This can be accomplished by issuing the following command: chdmod 777 ./turtleStart.sh when in the /jetsonbot/devel directory.